Struct deno_runtime::web_worker::WebWorker

pub struct WebWorker {
    pub js_runtime: JsRuntime,
    pub name: String,
    pub use_deno_namespace: bool,
    pub main_module: ModuleSpecifier,
    // some fields omitted
}

This struct is an implementation of Worker Web API

Each WebWorker is either a child of MainWorker or other WebWorker.

Fields
js_runtime: JsRuntimename: Stringuse_deno_namespace: boolmain_module: ModuleSpecifier
Implementations

impl WebWorker

pub fn from_options(
name: String,
permissions: Permissions,
main_module: ModuleSpecifier,
worker_id: WorkerId,
options: &WebWorkerOptions
) -> Self

pub fn bootstrap(&mut self, options: &WebWorkerOptions)

pub fn execute(&mut self, js_source: &str) -> Result<(), AnyError>

Same as execute2() but the filename defaults to “$CWD/anonymous”.

pub async fn preload_module(
&mut self,
module_specifier: &ModuleSpecifier
) -> Result<ModuleId, AnyError>

Loads and instantiates specified JavaScript module.

pub async fn execute_module(
&mut self,
module_specifier: &ModuleSpecifier
) -> Result<(), AnyError>

Loads, instantiates and executes specified JavaScript module.

pub fn thread_safe_handle(&self) -> WebWorkerHandle

Returns a way to communicate with the Worker from other threads.

pub fn poll_event_loop(
&mut self,
cx: &mut Context<'_>,
wait_for_inspector: bool
) -> Poll<Result<(), AnyError>>

pub async fn run_event_loop(
&mut self,
wait_for_inspector: bool
) -> Result<(), AnyError>